fix:post_idとqiitaのlogicとviewわけ
